move AddressDTO to correct folder
