+++ src/main/java/br/com/puppyplace/core/modules/customer/dto/CustomerDTO.java
@@ -1,8 +1,7 @@
 package br.com.puppyplace.core.modules.customer.dto;
 
+import br.com.puppyplace.core.modules.address.dto.AddressDTO;
 import com.fasterxml.jackson.annotation.JsonFormat;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.annotation.JsonProperty.Access;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
